clean up comments
